+++ src/test/java/org/sbrubbles/genericcons/MatchesVarargsTest.java
@@ -1,10 +1,7 @@
 package org.sbrubbles.genericcons;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -60,15 +57,9 @@ public class MatchesVarargsTest {
   @SuppressWarnings("unchecked")
   @Test
   public void nullTypesMultipleArgumentsMatch() {
-    try {
-      C.matchesVarargs(
-          Arrays.asList(Types.STRING, null, Types.OBJECT), 
-          null, new Double(1.0), null);
-      fail();
-    } catch (TypeMatchingException e) {
-      assertNull(e.getFullType());
-      assertEquals(new Double(1.0), e.getObject());
-    }
+    assertFalse(C.matchesVarargs(
+        Arrays.asList(Types.STRING, null, Types.OBJECT), 
+        null, new Double(1.0), null));
   }
   
   @SuppressWarnings("unchecked")